:recycle: api: move flash algo config to the right context
