+++ api/src/certification/configuration/domain/usecases/create-flash-assessment-configuration.js
