+++ api/src/certification/configuration/domain/usecases/get-active-flash-assessment-configuration.js
