+++ api/src/certification/configuration/infrastructure/repositories/flash-algorithm-configuration-repository.js
@@ -1,0 +1,6 @@
+import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
+
+export const save = async (flashAlgorithmConfiguration) => {
+  const knexConn = DomainTransaction.getConnection();
+  return knexConn('flash-algorithm-configurations').insert(flashAlgorithmConfiguration.toDTO());
+};
