+++ api/src/certification/configuration/routes.js
@@ -1,7 +1,9 @@
 import * as complementaryCertification from './application/complementary-certification-route.js';
+import * as flashAssessmentConfiguration from './application/flash-assessment-configuration-route.js';
 import * as scoWhitelist from './application/sco-whitelist-route.js';
 
 const certificationConfigurationRoutes = [complementaryCertification];
 const scoWhitelistRoutes = [scoWhitelist];
+const flashAssessmentConfigurationRoutes = [flashAssessmentConfiguration];
 
-export { certificationConfigurationRoutes, scoWhitelistRoutes };
+export { certificationConfigurationRoutes, flashAssessmentConfigurationRoutes, scoWhitelistRoutes };
