+++ api/src/certification/flash-certification/routes.js
@@ -1,6 +1,5 @@
-import * as flashAssessmentConfiguration from './application/flash-assessment-configuration-route.js';
 import * as scenarioSimulator from './application/scenario-simulator-route.js';
 
-const flashCertificationRoutes = [scenarioSimulator, flashAssessmentConfiguration];
+const flashCertificationRoutes = [scenarioSimulator];
 
 export { flashCertificationRoutes };
