+++ api/tests/certification/configuration/integration/application/flash-assessment-configuration-controller_test.js
@@ -1,5 +1,5 @@
-import { flashAssessmentConfigurationController } from '../../../../../src/certification/flash-certification/application/flash-assessment-configuration-controller.js';
-import { usecases } from '../../../../../src/certification/flash-certification/domain/usecases/index.js';
+import { flashAssessmentConfigurationController } from '../../../../../src/certification/configuration/application/flash-assessment-configuration-controller.js';
+import { usecases } from '../../../../../src/certification/configuration/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Integration | Application | FlashAssessmentConfigurationController', function () {
