+++ api/tests/certification/configuration/integration/infrastructure/repositories/flash-algorithm-configuration-repository_test.js
@@ -1,7 +1,7 @@
-import * as flashAlgorithmConfigurationRepository from '../../../../../../../api/src/certification/flash-certification/infrastructure/repositories/flash-algorithm-configuration-repository.js';
+import * as flashAlgorithmConfigurationRepository from '../../../../../../../api/src/certification/configuration/infrastructure/repositories/flash-algorithm-configuration-repository.js';
 import { domainBuilder, expect, knex } from '../../../../../test-helper.js';
 
-describe('Certification | Flash-certification | Integration | Infrastructure | Repository | FlashAlgorithmConfigurationRepository', function () {
+describe('Certification | configuration | Integration | Infrastructure | Repository | FlashAlgorithmConfigurationRepository', function () {
   describe('#save', function () {
     it('should create a flash algorithm configuration', async function () {
       // given
